engine: fix thread support for Linux
This allows enabling multithreading on Linux, without enabling it yet.
Linux uses pthread.h for threads, like Android, but in contrast to Android,
we have to define uIntArch ourselves.

On 64 bit Linux, that type has to be 64 bit wide and won't fit into
the uInt32 that was used in some places. Some type casts were added to
print statements, to avoid mismatches between the format string and
the actual uIntArch type.

A NULL was necessary because at one point, installOutput() was called
when there was no fDbgOutP, thus causing a crash on the
fDbgOutP->clone() call.

+++ src/sysync/stdlogicds.cpp
@@ -474,9 +474,9 @@ bool TStdLogicDS::threadedStartSync(void)
   PDEBUGPRINTFX(DBG_HOT,(
     "******* started &html;<a href=\"%s_%lu%s\" target=\"_blank\">&html;background thread id=%lu&html;</a>&html; for reading sync set",
     getDbgLogger()->getDebugFilename(), // href base
-    fStartSyncThread.getid(), // plus thread
+    (long unsigned)fStartSyncThread.getid(), // plus thread
     getDbgLogger()->getDebugExt(), // plus extension
-    fStartSyncThread.getid()
+    (long unsigned)fStartSyncThread.getid()
   ));
   #endif
   return true; // started ok
@@ -569,7 +569,7 @@ localstatus TStdLogicDS::startDataAccessForServer(void)
       if (fStartSyncThread.waitfor(t<0 ? 0 : t * 1000)) {
         // background thread has terminated
         sta = fStartSyncThread.exitcode();
-        PDEBUGPRINTFX(DBG_HOT,("******* background thread for startSync() terminated with exit code=%ld, status sta=%hd", fStartSyncThread.exitcode(),sta));
+        PDEBUGPRINTFX(DBG_HOT,("******* background thread for startSync() terminated with exit code=%ld, status sta=%hd", (long)fStartSyncThread.exitcode(),sta));
         // initialisation is now complete
         fInitializing=false;
       } // if
